requirements.txt
simplejson
Pillow
opencv-contrib-python
pytesseract
google-cloud-vision
beautifulsoup4
Vocabulary
lxml
Wikipedia-API